requirements.txt
PyGObject
toml
Pillow
pyscreenshot
